+++ test/function/statistics/max.test.js
@@ -78,6 +78,11 @@ describe('max', function() {
     assert.throws(function () {max(new Complex(3,4), 6)}, TypeError);
   });
 
+  it('should throw an error when called multiple arrays or matrices', function() {
+    assert.throws(function () {max([1,2], [3,4])}, /Scalar values expected/);
+    assert.throws(function () {max(math.matrix([1,2]), math.matrix([3,4]))}, /Scalar values expected/);
+  });
+
   it('should throw an error if called a dimension out of range', function() {
     assert.throws(function() {max([1,2,3], -1)}, /IndexError: Index out of range \(-1 < 0\)/);
     assert.throws(function() {max([1,2,3], 1)}, /IndexError: Index out of range \(1 > 0\)/);