+++ HISTORY.md
@@ -6,6 +6,8 @@
 - Fixed #621: a bug in parsing implicit multiplications like `(2)(3)+4`.
 - Fixed #623: `nthRoot` of zero with a negative root returned `0` instead of
   `Infinity`.
+- Throw an error when functions `min`, `max`, `mean`, or `median` are invoked
+  with multiple matrices as arguments (see #598).
 
 
 ## 2016-03-19, version 3.1.0